+++ ModRelease/CHANGELOG.md
@@ -15,16 +15,20 @@
 - Buffed Gup.
 - Buffed Geep.
 - Buffed Gip.
+- Berfed Glowing Meteorite Minimum Wave Interval 0.5s => **1s**.
+- Berfed Glowing Meteorite Maximum Wave Interval 1s => **1.5s**.
 - Nerfed Armor-Piercing Rounds Boss Damage Bonus 20% => **15%**.
 - Nerfed Leeching Seed to have a 0.15s cooldown between healing regardless of source.
 - Nerfed Spare Drone Parts Attack Speed Gain 50% => **40%**.
 - Nerfed Spare Drone Parts Cooldown Reduction 50% => **40%**.
+- Nerfed Spinel Tonic Post Processing.
 - Nerfed Stone Flux Pauldron pull on bosses.
 - Nerfed Elite Aspect Drop Chance 0.06% => **0.04%**.
 - Nerfed Wandering Vagrant.
 - Reworked Harvester's Scythe to swing a scythe on secondary skill use for 300% (+300% per stack) base damage, and if it hits, gain a 30% crit chance bonus for 4s.
 - Reworked Artifact of Command to drop potentials.
 - Increased Ally Cap 20 => 100.
+- Made Glowing Meteorite a Cylinder.
 - Made All Aspects apply on all allies without an equipment.
 - Made Gold Price Scaling decrease slightly per player in Multiplayer.
 - Made Printer Common and Uncommon Max Uses scale with player count.